feat: add decodeFromReader
Instead of pulling in an entire chunk, that may consist of multiple messages,
decodeFromReader is able to read exactly the number of bytes from the reader
as needed. This helps prevent overreading when you have access to the reader.

This is particularly useful in situations where you need to perform a handshake
and then continue streaming data once the handshake is complete.