fix!: do not create unecessary async work
When a synchronous iterator is passed, return a synchronous generator.

To get the types to play nicely it was necessary to change the function signature:

Before:

```js
pipe(
  byteStream,
  lp.decode(),
  async function * (messages) {
    // ..yield messages here
  },
  lp.decode(),
  byteStream
)
```

After:

```js
pipe(
  byteStream,
  (source) => lp.decode(source),
  async function * (messages) {
    // ..yield messages here
  },
  (source) => lp.decode(source),
  byteStream
)
```

This is because typescript cannot derive the return type of the returned function based on the
input without being aware of the input type.

BREAKING CHANGE: a synchronous generator will be returned when a synchronous iterator is passed